refactor: Move AmaiSosu.Installation.OpenSauceException to AmaiSosu.Common namespace
